pyproject.toml
[project]
name = "botreformed-baptist-1689-bible-expert"
version = "0.1.0"
description = ""
requires-python = ">=3.10,<4.0"
dependencies = [
    "mlserver==1.5.0",
    "huggingface-hub~=0.34.4",
    "numpy~=2.2.6",
    "llama-cpp-python~=0.3.15",
]

[tool.uv]
dev-dependencies = [
    "pytest<8.0.0,>=7.4.3",
    "mypy<2.0.0,>=1.8.0",
    "ruff<1.0.0,>=0.2.2",
    "pre-commit<4.0.0,>=3.6.2",
    "types-passlib<2.0.0.0,>=1.7.7.20240106",
    "coverage<8.0.0,>=7.4.3",
    "black>=25.1.0",
    "behave~=1.3.1",
]

[build-system]
requires = [
    "hatchling"
]

build-backend = "hatchling.build"

[tool.mypy]
strict = true
exclude = ["venv", ".venv"]

[tool.ruff]
target-version = "py310"
exclude = []

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "ARG001", # unused arguments in functions
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "W191",  # indentation contains tabs
    "B904",  # Allow raising exceptions without from e, for HTTPException
]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true


# Behave configuration
[tool.behave]
paths = ["features"]
format = ["pretty", "json.pretty", "junit", "html"]
show_timings = true
show_multiline = true
junit = true
junit_directory = "reports/junit"
json = "reports/json/report.json"
html = "reports/html/report.html"
logging_level = "INFO"

[tool.behave.userdata]
mlserver_url = "http://localhost:8080"
model_name = "reformed-baptist-bible-expert"
timeout = 30